Use the head branch's touchstone script.
When setting up branches, we were switching from the head to the base
branch. This implies that the benchmarking script being used is that of
the base branch.

I believe it makes more sense to use the head branch's script, as when
one adds a new benchmark to an existing script, you would want to see to
results of it without having to merge that PR first.

This actually used to be the behaviour but was changed as part of #112.
I cannot see any discussion about this change so I suspect it may have
been unintentional.