Fix how the author association is looked up for PRs.
The use_touchstone_workflows was generating code that used the
`github.event.comment.author_association` regardless of the trigger
type. However this does not work for pull_request triggers, meaning the
if condition would always fail to evaluate and the workflow never ran.
This commit fixes that by using the correct field name.