Upload raw benchmark results as an artifact.
The raw benchmark results weren't available anywhere. Exposing them as
an artifact is convenient for investigation and to customize the
rendering and visualization after the fact.

The receive action was uploading the PR comment as an artifact twice,
once named "pr" and once named "results". Only the former is used by the
comment action. Given the redundancy I assume this must have been a
mistake. The "results" artifact now contains the raw benchmarking
numbers instead.

I've also removed a spurious download-artifact step, which was
downloading the same files that had just been uploaded, for no apparent
reason.